create a dedicated sound service
