+++ src/services/index.js
@@ -1,3 +1,4 @@
+import './sound'
 import './speech'
 import './storage'
 // lint .-.
